requirements.txt
Django==2.0.7
djangorestframework==3.8.2
gunicorn==19.9.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
